requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
tweepy==4.14.0
schedule==1.2.0
python-dotenv==1.0.0
pytz==2023.3
lxml==4.9.3
feedparser==6.0.10
newspaper3k==0.2.8
nltk==3.8.1
textblob==0.17.1
pandas==2.1.3
python-dateutil==2.8.2
webdriver-manager==4.0.1

